requirements.txt
# Core LLM Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Interface
streamlit>=1.28.0

# OpenAI API
openai>=1.0.0

# HuggingFace Transformers

# Environment Management
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# HTTP Requests
requests>=2.31.0

# Text Processing
tiktoken>=0.5.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Data visualization (often needed for analysis)
matplotlib>=3.7.0
seaborn>=0.12.0

# Excel file handling (since you have Excel files in your project)
openpyxl>=3.1.0
xlrd>=2.0.0

# JSON handling for API responses
jsonschema>=4.17.0

# Progress bars and logging
tqdm>=4.65.0
loguru>=0.7.0

# Vector databases (commonly used with LLMs)
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Text embeddings
sentence-transformers>=2.2.0 